Split import wizard step cards
